+++ etcdctl/ctlv3/command/printer.go
@@ -21,7 +21,7 @@ import (
 
 	pb "go.etcd.io/etcd/api/v3/etcdserverpb"
 	v3 "go.etcd.io/etcd/v3/clientv3"
-	"go.etcd.io/etcd/v3/clientv3/snapshot"
+	"go.etcd.io/etcd/v3/etcdctl/snapshot"
 
 	"github.com/dustin/go-humanize"
 )
