+++ etcdctl/ctlv3/command/printer_json.go
@@ -22,7 +22,7 @@ import (
 	"strconv"
 
 	"go.etcd.io/etcd/v3/clientv3"
-	"go.etcd.io/etcd/v3/clientv3/snapshot"
+	"go.etcd.io/etcd/v3/etcdctl/snapshot"
 )
 
 type jsonPrinter struct {
