+++ etcdctl/ctlv3/command/printer_table.go
@@ -18,7 +18,7 @@ import (
 	"os"
 
 	v3 "go.etcd.io/etcd/v3/clientv3"
-	"go.etcd.io/etcd/v3/clientv3/snapshot"
+	"go.etcd.io/etcd/v3/etcdctl/snapshot"
 
 	"github.com/olekukonko/tablewriter"
 )
